Fix another 64 bit issue
git-svn-id: https://qwaveclient.googlecode.com/svn/trunk@258 4cf79858-b814-11de-b7b1-c13d6a5ff25b

+++ waveserver/fcgi/fcgiclientconnection.cpp
@@ -71,7 +71,7 @@ void FCGIClientConnection::customEvent( QEvent* event )
 
 void FCGIClientConnection::handleRequest( const PBMessage<webclient::Request>* request )
 {
-    m_ackedSequenceNumber = qMax( request->client_ack(), m_ackedSequenceNumber );
+    m_ackedSequenceNumber = qMax( (qint64)request->client_ack(), m_ackedSequenceNumber );
     // TODO: If the client sequence number is not what we expected?
     m_clientSequenceNumber = request->client_sequence_number();
